pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "epgb-pyrofex"
version = "1.0.0"
description = "EPGB Options Market Data - Python market data fetcher with Excel integration"
readme = "README.md"
authors = [
    {name = "EPGB Options", email = "developer@epgb.com"}
]
requires-python = ">=3.9"

# Core dependencies
dependencies = [
    "pyRofex>=0.5.0",
    "xlwings>=0.31.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.8.0",
]

# Entry points for command-line scripts
[project.scripts]
epgb-options = "epgb_options.main:main"

# Setuptools configuration
[tool.setuptools.packages.find]
where = ["src"]